fix(primitives): make portal attachment reactive
