+++ packages/components/src/components/focus-trap/focus-trap.test.ts
@@ -8,7 +8,6 @@ setupHappyDom();
 
 const { render, fireEvent, cleanup } = await import('@testing-library/svelte');
 const { default: FocusTrap } = await import('./focus-trap.svelte');
-const { createFocusTrap } = await import('./focus-trap.utilities.ts');
 
 const focusTrapChildren = createRawSnippet(() => ({
   render: () => `
@@ -49,6 +48,7 @@ describe('FocusTrap', () => {
 
     const first = getByTestId('first-button') as HTMLButtonElement;
     const last = getByTestId('last-button') as HTMLButtonElement;
+    await tick();
     last.focus();
 
     await fireEvent.keyDown(last, { key: 'Tab' });
@@ -56,25 +56,19 @@ describe('FocusTrap', () => {
     expect(document.activeElement).toBe(first);
   });
 
-  test('wraps Shift+Tab from the first element to the last', () => {
-    const root = document.createElement('div');
-    root.innerHTML = `
-      <button data-testid="first-button">First</button>
-      <button data-testid="last-button">Last</button>
-    `;
-    document.body.appendChild(root);
-    const detach = createFocusTrap()(root);
+  test('wraps Shift+Tab from the first element to the last', async () => {
+    const { getByTestId } = render(FocusTrap, {
+      props: { children: focusTrapChildren },
+    });
 
-    const first = root.querySelector('[data-testid="first-button"]') as HTMLButtonElement;
-    const last = root.querySelector('[data-testid="last-button"]') as HTMLButtonElement;
+    const first = getByTestId('first-button') as HTMLButtonElement;
+    const last = getByTestId('last-button') as HTMLButtonElement;
+    await tick();
     first.focus();
 
-    first.dispatchEvent(
-      new KeyboardEvent('keydown', { key: 'Tab', shiftKey: true, bubbles: true }),
-    );
+    await fireEvent.keyDown(first, { key: 'Tab', shiftKey: true, bubbles: true });
 
     expect(document.activeElement).toBe(last);
-    detach?.();
   });
 
   test('restores focus to the previously focused element on teardown', async () => {
