+++ packages/components/src/components/focus-trap/index.ts
@@ -2,5 +2,5 @@ import FocusTrap from './focus-trap.svelte';
 
 export default FocusTrap;
 export type { FocusTrapOptions, FocusTrapProps } from './focus-trap.types.ts';
-export { createFocusTrap } from './focus-trap.utilities.ts';
+export { createFocusTrap } from './focus-trap.utilities.svelte.ts';
 export { FocusTrap };
