+++ packages/components/src/components/portal/index.ts
@@ -2,5 +2,5 @@ import Portal from './portal.svelte';
 
 export default Portal;
 export type { PortalAttachmentOptions, PortalProps } from './portal.types.ts';
-export { createPortalAttachment } from './portal.utilities.ts';
+export { createPortalAttachment } from './portal.utilities.svelte.ts';
 export { Portal };
