+++ packages/components/src/components/portal/portal.test.ts
@@ -80,4 +80,44 @@ describe('Portal', () => {
     expect(result.ssrHtml).not.toContain('Portaled child');
     result.cleanup();
   });
+
+  test('retargets when the target prop changes after mount', async () => {
+    const hostA = document.createElement('div');
+    hostA.id = 'portal-host-a';
+    const hostB = document.createElement('div');
+    hostB.id = 'portal-host-b';
+    document.body.append(hostA, hostB);
+
+    const { rerender } = render(Portal, {
+      props: { target: '#portal-host-a', children: childSnippet },
+    });
+
+    await tick();
+    expect(hostA.querySelector('[data-testid="portal-child"]')).not.toBeNull();
+
+    await rerender({ target: '#portal-host-b', children: childSnippet });
+    await tick();
+
+    expect(hostA.querySelector('[data-testid="portal-child"]')).toBeNull();
+    expect(hostB.querySelector('[data-testid="portal-child"]')).not.toBeNull();
+  });
+
+  test('detaches when disabled flips from false to true', async () => {
+    const host = document.createElement('div');
+    host.id = 'portal-host';
+    document.body.appendChild(host);
+
+    const { rerender } = render(Portal, {
+      props: { target: '#portal-host', disabled: false, children: childSnippet },
+    });
+
+    await tick();
+    expect(host.querySelector('[data-testid="portal-child"]')).not.toBeNull();
+
+    await rerender({ target: '#portal-host', disabled: true, children: childSnippet });
+    await tick();
+
+    // After disabling, the wrapper must not remain in the previous target.
+    expect(host.querySelector('[data-testid="portal-child"]')).toBeNull();
+  });
 });
